rename RQModal to RQFormCard
